+++ src/common/dto/utils/index.ts
@@ -1,1 +1,2 @@
 export { default as IsMutuallyExclusiveWith } from './IsMutuallyExclusiveWith';
+export * from './transformers';
